# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Goat Farm Management API",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/feeding/plan": {
      "get": {
        "tags": [
          "feeding"
        ],
        "summary": "Feeding Today",
        "description": "Today's 3-shift plan (bucket \u00d7 recipe \u00d7 headcount) + today's dispensing log.",
        "operationId": "feeding_today_api_feeding_plan_get",
        "parameters": [
          {
            "name": "x-farm-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Farm-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedingPlanOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/feeding/settings": {
      "post": {
        "tags": [
          "feeding"
        ],
        "summary": "Save Setting",
        "description": "Override the per-head daily ration (kg) for one bucket.",
        "operationId": "save_setting_api_feeding_settings_post",
        "parameters": [
          {
            "name": "x-farm-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Farm-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedSettingIn"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/feeding/dispense": {
      "post": {
        "tags": [
          "feeding"
        ],
        "summary": "Dispense",
        "description": "Record feed actually dispensed to a bucket on a shift.",
        "operationId": "dispense_api_feeding_dispense_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-farm-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Farm-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispenseIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedingRecordOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/feeding/recipes": {
      "get": {
        "tags": [
          "feeding"
        ],
        "summary": "List Recipes",
        "description": "All feed recipes (with lines) + the bucket allocation reference table.",
        "operationId": "list_recipes_api_feeding_recipes_get",
        "parameters": [
          {
            "name": "x-farm-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Farm-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/feeding/mix": {
      "post": {
        "tags": [
          "feeding"
        ],
        "summary": "Mix Batch",
        "description": "Mix a batch of a recipe, decrementing inventory per recipe lines.",
        "operationId": "mix_batch_api_feeding_mix_post",
        "parameters": [
          {
            "name": "x-farm-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Farm-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MixIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedRecipeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/feeding/inventory": {
      "get": {
        "tags": [
          "feeding"
        ],
        "summary": "List Inventory",
        "description": "Feed inventory for this farm, ordered by category then ingredient.",
        "operationId": "list_inventory_api_feeding_inventory_get",
        "parameters": [
          {
            "name": "x-farm-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Farm-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeedInventoryOut"
                  },
                  "title": "Response List Inventory Api Feeding Inventory Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/feeding/inventory/{item_id}/add": {
      "post": {
        "tags": [
          "feeding"
        ],
        "summary": "Add Stock",
        "description": "Purchase stock: bump qty on hand, update last price, book a FEED expense.",
        "operationId": "add_stock_api_feeding_inventory__item_id__add_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-farm-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Farm-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockAddIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedInventoryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BucketAllocationOut": {
        "properties": {
          "bucket": {
            "type": "string",
            "title": "Bucket"
          },
          "allocation": {
            "type": "string",
            "title": "Allocation"
          }
        },
        "type": "object",
        "required": [
          "bucket",
          "allocation"
        ],
        "title": "BucketAllocationOut"
      },
...
      "DispenseIn": {
        "properties": {
          "bucket": {
            "type": "string",
            "enum": [
              "QUARANTINE",
              "FOUNDATION",
              "BREEDING",
              "PREGNANCY_EARLY",
              "PREGNANCY_LATE",
              "DELIVERY",
              "RECOVERY",
              "RESTING",
              "MALE_KIDS",
              "FEMALE_KIDS"
            ],
            "title": "Bucket"
          },
          "shift": {
            "type": "string",
            "enum": [
              "MORNING",
              "AFTERNOON",
              "NIGHT"
            ],
            "title": "Shift"
          },
          "recipe_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipe Code"
          },
          "qty_kg": {
            "type": "number",
            "title": "Qty Kg",
            "le": 1000000
          },
          "date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          }
        },
        "type": "object",
        "required": [
          "bucket",
          "shift",
          "qty_kg"
        ],
        "title": "DispenseIn"
      },
...
      "FeedInventoryOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "ingredient": {
            "type": "string",
            "title": "Ingredient"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "qty_on_hand": {
            "type": "number",
            "title": "Qty On Hand"
          },
          "reorder_level": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reorder Level"
          },
          "last_purchase_price_per_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Purchase Price Per Kg"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ingredient",
          "category",
          "unit",
          "qty_on_hand",
          "reorder_level",
          "last_purchase_price_per_kg"
        ],
        "title": "FeedInventoryOut"
      },
      "FeedRecipeLineOut": {
        "properties": {
          "ingredient": {
            "type": "string",
            "title": "Ingredient"
          },
          "kg_per_100kg": {
            "type": "number",
            "title": "Kg Per 100Kg"
          },
          "category": {
            "type": "string",
            "title": "Category"
          }
        },
        "type": "object",
        "required": [
          "ingredient",
          "kg_per_100kg",
          "category"
        ],
        "title": "FeedRecipeLineOut"
      },
      "FeedRecipeOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/FeedRecipeLineOut"
            },
            "type": "array",
            "title": "Lines",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "description"
        ],
        "title": "FeedRecipeOut"
      },
      "FeedSettingIn": {
        "properties": {
          "bucket": {
            "type": "string",
            "enum": [
              "QUARANTINE",
              "FOUNDATION",
              "BREEDING",
              "PREGNANCY_EARLY",
              "PREGNANCY_LATE",
              "DELIVERY",
              "RECOVERY",
              "RESTING",
              "MALE_KIDS",
              "FEMALE_KIDS"
            ],
            "title": "Bucket"
          },
          "daily_kg_per_head": {
            "type": "number",
            "title": "Daily Kg Per Head",
            "le": 1000000
          }
        },
        "type": "object",
        "required": [
          "bucket",
          "daily_kg_per_head"
        ],
        "title": "FeedSettingIn"
      },
...
      "FeedingPlanOut": {
        "properties": {
          "lines": {
            "items": {
              "$ref": "#/components/schemas/PlanLineOut"
            },
            "type": "array",
            "title": "Lines"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/FeedingRecordOut"
            },
            "type": "array",
            "title": "Records"
          }
        },
        "type": "object",
        "required": [
          "lines",
          "records"
        ],
        "title": "FeedingPlanOut"
      },
      "FeedingRecordOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "shift": {
            "type": "string",
            "title": "Shift"
          },
          "bucket": {
            "type": "string",
            "title": "Bucket"
          },
          "recipe_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipe Code"
          },
          "qty_kg": {
            "type": "number",
            "title": "Qty Kg"
          }
        },
        "type": "object",
        "required": [
          "id",
          "date",
          "shift",
          "bucket",
          "recipe_code",
          "qty_kg"
        ],
        "title": "FeedingRecordOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MixIn": {
        "properties": {
          "recipe_code": {
            "type": "string",
            "minLength": 1,
            "title": "Recipe Code"
          },
          "batch_kg": {
            "type": "number",
            "title": "Batch Kg",
            "le": 1000000
          }
        },
        "type": "object",
        "required": [
          "recipe_code",
          "batch_kg"
        ],
        "title": "MixIn"
      },
...
      "PlanLineOut": {
        "properties": {
          "bucket": {
            "type": "string",
            "title": "Bucket"
          },
          "recipe_code": {
            "type": "string",
            "title": "Recipe Code"
          },
          "recipe_name": {
            "type": "string",
            "title": "Recipe Name"
          },
          "heads": {
            "type": "integer",
            "title": "Heads"
          },
          "kg_per_head": {
            "type": "number",
            "title": "Kg Per Head"
          },
          "daily_kg": {
            "type": "number",
            "title": "Daily Kg"
          },
          "shifts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Shifts"
          }
        },
        "type": "object",
        "required": [
          "bucket",
          "recipe_code",
          "recipe_name",
          "heads",
          "kg_per_head",
          "daily_kg",
          "shifts"
        ],
        "title": "PlanLineOut"
      },
...
      "RecipeListOut": {
        "properties": {
          "recipes": {
            "items": {
              "$ref": "#/components/schemas/FeedRecipeOut"
            },
            "type": "array",
            "title": "Recipes"
          },
          "allocation": {
            "items": {
              "$ref": "#/components/schemas/BucketAllocationOut"
            },
            "type": "array",
            "title": "Allocation"
          }
        },
        "type": "object",
        "required": [
          "recipes",
          "allocation"
        ],
        "title": "RecipeListOut"
      },
...
      "StockAddIn": {
        "properties": {
          "qty_kg": {
            "type": "number",
            "title": "Qty Kg",
            "le": 1000000
          },
          "price_per_kg": {
            "anyOf": [
              {
                "type": "number",
                "le": 1000000000
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Per Kg"
          }
        },
        "type": "object",
        "required": [
          "qty_kg"
        ],
        "title": "StockAddIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}